# CDC Optimization: Don't use Get in newChunkedReader